watchlater_likedvideos_and_comment_ui

+++ client/src/components/Comment/DisplayComment.jsx
@@ -1,7 +1,10 @@
 import moment from "moment";
-import React, { useEffect, useState } from "react";
+import React, { useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { deleteComment, editComment } from "../../redux/action/comment";
+import { BsThreeDotsVertical } from "react-icons/bs";
+import { MdOutlineEdit } from "react-icons/md";
+import { RiDeleteBin6Line } from "react-icons/ri";
 
 function DisplayComment({
   cId,
@@ -13,35 +16,44 @@ function DisplayComment({
   const [edit, setEdit] = useState(false);
   const [cmtId, setCmtId] = useState("");
   const [cmtBody, setCmtBody] = useState("");
-  const dispatch = useDispatch()
-  const currentUser = useSelector(state=>state.currentUserReducer);
-  
+  const [cmtMenu,setCmtMenu] = useState(false)
+  const dispatch = useDispatch();
+  const currentUser = useSelector((state) => state.currentUserReducer);
+
   const handleEdit = (cId, cBody) => {
+    setCmtMenu(false)
     setEdit(true);
     setCmtId(cId);
     setCmtBody(cBody);
   };
 
-  const handleOnSubmit = (e)=>{
+  const handleOnSubmit = (e) => {
     e.preventDefault();
-    if(!cmtBody){
-      alert("type your comment!")
-    }else{
-      dispatch(editComment({id:cmtId,commentBody:cmtBody}))
-      setCmtBody("")
+    if (!cmtBody) {
+      alert("type your comment!");
+    } else {
+      dispatch(editComment({ id: cmtId, commentBody: cmtBody }));
+      setCmtBody("");
     }
-    setEdit(false)
-  }
+    setEdit(false);
+  };
 
-  const handleDelete = (id) =>{
-    dispatch(deleteComment(id))
-  }
+  const handleDelete = (id) => {
+    const confirmDel = window.confirm("Delete your comment permanently?")
+    if(confirmDel){
+      dispatch(deleteComment(id))
+    }
+    setCmtMenu(false)
+  };
 
   return (
-    <>
+    <div id="container_main_comment">
       {edit ? (
         <>
-          <form className="comments_sub_form_comments" onSubmit={(e)=>handleOnSubmit(e)}>
+          <form
+            className="comments_sub_form_comments"
+            onSubmit={(e) => handleOnSubmit(e)}
+          >
             <input
               type="text"
               onChange={(e) => setCmtBody(e.target.value)}
@@ -54,17 +66,44 @@ function DisplayComment({
               value="Change"
               className="comment_add_btn_comments"
             />
+            <input className="comment_add_btn_comments" value="Cancel" readOnly onClick={() => setEdit(false)}/>
           </form>
         </>
       ) : (
-          <>
-          <p className="usercommented">{""} @{userCommented?.toLowerCase()} <span className="commenton">commented {moment(commentOn).fromNow()}</span></p>
-          <p className="comment_body">{commentBody}</p>
+        <div className="comment_container">
+          <div>
+            <div className="Chanel_logo_App" id="comment_channel_logo">
+              <p className="fstChar_logo_App">
+                <>{userCommented.charAt(0).toUpperCase()}</>
+              </p>
+            </div>
+          </div>
+          <div className="comment_content">
+            <p className="usercommented">
+              {""} @{userCommented?.toLowerCase()}{" "}
+              <span className="commenton">
+                commented {moment(commentOn).fromNow()}
+              </span>
+            </p>
+            <p className="comment_body">{commentBody}</p>
+          </div>
+        </div>
+      )}
+      {currentUser?.result?._id === userId && (
+        <>
+        <p className="comment_options" onClick={()=>setCmtMenu(!cmtMenu)}><BsThreeDotsVertical /></p>
+        { cmtMenu?  
+        <p className="EditDel_DisplayComment">
+          <span onClick={() => handleEdit(cId, commentBody)}><MdOutlineEdit/>&nbsp; Edit</span>
+          <span onClick={() => handleDelete(cId)}><RiDeleteBin6Line/>&nbsp; Delete</span>
+        </p>:""
+        }
+        {
+          cmtMenu? <div className="menu_bg" onClick={()=>setCmtMenu(false)}></div> : ""
+        }
         </>
       )}
-      {currentUser?.result?._id===userId&&(<p className="EditDel_DisplayComment"><i onClick={()=>handleEdit(cId,commentBody)}>Edit</i>&nbsp;&nbsp;&nbsp;<i onClick={()=>handleDelete(cId)}>Delete</i></p>)}
-      
-    </>
+    </div>
   );
 }
